Only remove leftover echo file if it exists.

+++ casper_extension_attributes/shellshock_check.py
@@ -80,7 +80,8 @@ def main():
 
     if len(bad_shells) > 0:
         print('<result>%s</result>' % ', '.join(bad_shells))
-        os.remove(os.path.join(os.getcwd(), 'echo'))
+        if os.path.exists(os.path.join(os.getcwd(), 'echo')):
+            os.remove(os.path.join(os.getcwd(), 'echo'))
     else:
         print('<result>Patched</result>')
